Added title to product hero block
